requirements.txt
numpy
opencv-python
Pillow 
Flask
tensorflow
keras
Werkzeug